Introduce Long as type for all ids (primary keys)

+++ src/main/java/org/synyx/urlaubsverwaltung/absence/AbsencePeriod.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 /**
  * Defines absence periods of a {@link Person}.
- *
+ * <p>
  * e.g. Bruce Wayne is absent on:
  * <ul>
  *     <li>24.March 2021 to 28. March 2021 (vacation full day)</li>
@@ -157,26 +157,37 @@ public class AbsencePeriod {
      */
     public interface RecordInfo {
         Person getPerson();
+
         AbsenceType getType();
+
         AbsenceStatus getStatus();
-        Integer getId();
+
+        Long getId();
+
         boolean hasStatusTemporaryAllowed();
+
         boolean hasStatusWaiting();
+
         boolean hasStatusAllowed();
+
         boolean hasStatusAllowedCancellationRequested();
-        Optional<Integer> getVacationTypeId();
+
+        Optional<Long> getVacationTypeId();
+
         boolean isVisibleToEveryone();
     }
 
     /**
      * Describes an absence this morning.
      */
-    public interface RecordMorning extends RecordInfo {}
+    public interface RecordMorning extends RecordInfo {
+    }
 
     /**
      * Describes an absence this noon.
      */
-    public interface RecordNoon extends RecordInfo {}
+    public interface RecordNoon extends RecordInfo {
+    }
 
     /**
      * Describes an absence record. (e.g. morning absence or noon absence)
@@ -185,17 +196,17 @@ public class AbsencePeriod {
 
         private final Person person;
         private final AbsenceType type;
-        private final Integer id;
+        private final Long id;
         private final AbsenceStatus status;
-        private final Integer vacationTypeId;
+        private final Long vacationTypeId;
 
         private final boolean visibleToEveryone;
 
-        private AbstractRecordInfo(Person person, AbsenceType type, Integer id, AbsenceStatus status) {
+        private AbstractRecordInfo(Person person, AbsenceType type, Long id, AbsenceStatus status) {
             this(person, type, id, status, null, false);
         }
 
-        private AbstractRecordInfo(Person person, AbsenceType type, Integer id, AbsenceStatus status, Integer vacationTypeId, boolean visibleToEveryone) {
+        private AbstractRecordInfo(Person person, AbsenceType type, Long id, AbsenceStatus status, Long vacationTypeId, boolean visibleToEveryone) {
             this.person = person;
             this.type = type;
             this.id = id;
@@ -215,7 +226,7 @@ public class AbsencePeriod {
         }
 
         @Override
-        public Integer getId() {
+        public Long getId() {
             return id;
         }
 
@@ -244,7 +255,7 @@ public class AbsencePeriod {
             return hasStatusOneOf(AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED);
         }
 
-        public Optional<Integer> getVacationTypeId() {
+        public Optional<Long> getVacationTypeId() {
             return Optional.ofNullable(vacationTypeId);
         }
 
@@ -274,25 +285,25 @@ public class AbsencePeriod {
     }
 
     public static class RecordMorningVacation extends AbstractRecordInfo implements RecordMorning {
-        public RecordMorningVacation(Person person, Integer applicationId, AbsenceStatus status, Integer vacationTypeId, boolean visibleToEveryone) {
+        public RecordMorningVacation(Person person, Long applicationId, AbsenceStatus status, Long vacationTypeId, boolean visibleToEveryone) {
             super(person, AbsenceType.VACATION, applicationId, status, vacationTypeId, visibleToEveryone);
         }
     }
 
     public static class RecordMorningSick extends AbstractRecordInfo implements RecordMorning {
-        public RecordMorningSick(Person person, Integer sickNoteId, AbsenceStatus status) {
+        public RecordMorningSick(Person person, Long sickNoteId, AbsenceStatus status) {
             super(person, AbsenceType.SICK, sickNoteId, status);
         }
     }
 
     public static class RecordNoonVacation extends AbstractRecordInfo implements RecordNoon {
-        public RecordNoonVacation(Person person, Integer applicationId, AbsenceStatus status, Integer vacationTypeId, boolean visibleToEveryone) {
+        public RecordNoonVacation(Person person, Long applicationId, AbsenceStatus status, Long vacationTypeId, boolean visibleToEveryone) {
             super(person, AbsenceType.VACATION, applicationId, status, vacationTypeId, visibleToEveryone);
         }
     }
 
     public static class RecordNoonSick extends AbstractRecordInfo implements RecordNoon {
-        public RecordNoonSick(Person person, Integer sickNoteId, AbsenceStatus status) {
+        public RecordNoonSick(Person person, Long sickNoteId, AbsenceStatus status) {
             super(person, AbsenceType.SICK, sickNoteId, status);
         }
     }
