+++ src/main/java/org/synyx/urlaubsverwaltung/application/application/ApplicationForLeaveViewController.java
@@ -195,7 +195,7 @@ class ApplicationForLeaveViewController implements HasLaunchpad {
     }
 
     private static ApplicationPersonDto toViewPerson(Person person, boolean allowedToAccessPersonData) {
-        final Integer id = allowedToAccessPersonData ? person.getId() : null;
+        final Long id = allowedToAccessPersonData ? person.getId() : null;
         return new ApplicationPersonDto(person.getNiceName(), person.getGravatarURL(), person.isInactive(), id);
     }
