+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/AbsenceMappingServiceImpl.java
@@ -17,7 +17,7 @@ public class AbsenceMappingServiceImpl implements AbsenceMappingService {
     }
 
     @Override
-    public AbsenceMapping create(Integer id, AbsenceMappingType absenceMappingType, String eventId) {
+    public AbsenceMapping create(Long id, AbsenceMappingType absenceMappingType, String eventId) {
         final AbsenceMapping absenceMapping = new AbsenceMapping(id, absenceMappingType, eventId);
         return absenceMappingRepository.save(absenceMapping);
     }
@@ -28,7 +28,7 @@ public class AbsenceMappingServiceImpl implements AbsenceMappingService {
     }
 
     @Override
-    public Optional<AbsenceMapping> getAbsenceByIdAndType(Integer id, AbsenceMappingType absenceMappingType) {
+    public Optional<AbsenceMapping> getAbsenceByIdAndType(Long id, AbsenceMappingType absenceMappingType) {
         return absenceMappingRepository.findAbsenceMappingByAbsenceIdAndAbsenceMappingType(id, absenceMappingType);
     }
 }
