+++ src/main/java/org/synyx/urlaubsverwaltung/overview/OverviewViewController.java
@@ -103,7 +103,7 @@ public class OverviewViewController implements HasLaunchpad {
     }
 
     @GetMapping("/web/person/{personId}/overview")
-    public String showOverview(@PathVariable("personId") Integer personId,
+    public String showOverview(@PathVariable("personId") Long personId,
                                @RequestParam(value = "year", required = false) Integer year, Model model)
         throws UnknownPersonException {
