+++ src/main/java/org/synyx/urlaubsverwaltung/settings/AbsenceTypeSettingsItemDto.java
@@ -7,7 +7,7 @@ import java.util.Objects;
 
 public class AbsenceTypeSettingsItemDto {
 
-    private Integer id;
+    private Long id;
     private boolean active;
     private String messageKey;
     private VacationCategory category;
@@ -16,11 +16,11 @@ public class AbsenceTypeSettingsItemDto {
     private VacationTypeColor color;
     private boolean visibleToEveryone;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -119,7 +119,7 @@ public class AbsenceTypeSettingsItemDto {
     }
 
     static class Builder {
-        private Integer id;
+        private Long id;
         private boolean active;
         private String messageKey;
         private VacationCategory category;
@@ -128,7 +128,7 @@ public class AbsenceTypeSettingsItemDto {
         private VacationTypeColor color;
         private boolean visibleToEveryone;
 
-        Builder setId(Integer id) {
+        Builder setId(Long id) {
             this.id = id;
             return this;
         }
